Replace `@pytest.fixture` with `@pytest_asyncio.fixture`
This should fix a pytest warning

`/miniforge3/envs/jupyter/lib/python3.13/site-packages/_pytest/fixtures.py:1181: PytestRemovedIn9Warning: 'test_' requested an async fixture 'configured_mockhub_instance', with no plugin or hook that handled it. This is usually an error, as pytest does not natively support it. This will turn into an error in pytest 9.
  See: https://docs.pytest.org/en/stable/deprecations.html#sync-test-depending-on-async-fixture
`